+++ src/lua5.1/loslib.c
@@ -60,7 +60,7 @@ static int os_execute (lua_State *L) {
 
 /* converts UTF-8 string to UTF-16 string, returns also Win32 error */
 static wchar_t *convert_to_utf16(const char *utf8_string,int *error) {
-  int length = strlen(utf8_string);
+  int length = (int)strlen(utf8_string);
   wchar_t *utf16_string = malloc((length + 1) * 2);
   if(utf16_string == NULL){
     *error = ERROR_NOT_ENOUGH_MEMORY;
@@ -76,7 +76,7 @@ static wchar_t *convert_to_utf16(const char *utf8_string,int *error) {
 
 /* converts UTF-16 string to UTF-8 string, returns also Win32 error */
 static char *convert_to_utf8(const wchar_t *utf16_string,int *error) {
-  int length = wcslen(utf16_string);
+  int length = (int)wcslen(utf16_string);
   char *utf8_string = malloc((length + 1) * 2);
   if(utf8_string == NULL){
     *error = ERROR_NOT_ENOUGH_MEMORY;
